requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.3
dj-database-url==0.5.0
Django==2.2.5
djangorestframework==3.10.3
freeze==1.0.10
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
psycopg2==2.8.4
pybreaker==0.6.0
python-dateutil==1.5
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.3
